# Optimize ops.triangular_solve() for 1x1 matrices